Rename correct minified backup folder.
Minified backup must be extracted to the folder corresponding to
the master backup environment. In the current state it always
expects it to be production. As a result, after extracting the
backup, it tries to rename the `productionMinified` folder, which
in certain circumstances might not be the case.

Finally, if the backup master system is configured as DEPLOY, then
uncompressed minified backup folder should be deployMinified and
it should be renamed to deploy.